requirements/base.txt
Django==1.10.5
dj-database-url==0.4.1
django-model-utils==2.6.0
django-cors-headers==1.3.0
django-avatar==3.1.0
django-filter==0.15.3
django-crispy-forms==1.6.1
easy-thumbnails==2.3
django-fsm==2.4.0
django-fsm-admin==1.2.4
python-social-auth==0.2.21
django-push-notifications==1.4.1

# sending emails
djmail==0.13.0

# Django REST Framework
djangorestframework==3.5.3
djangorestframework-expander==0.2.3
djangorestframework-gis==0.11
git+https://github.com/silverlogic/drf-extra-fields.git@c5845d4de9873950d65bad3f487a6dc62f544e08
git+https://github.com/silverlogic/django-rest-social-auth@master

# postgresql driver
psycopg2==2.6.2

# runs the app server
gunicorn==19.6.0

# datetime helpers
python-dateutil==2.6.0

# bitcoin
coinbase==2.0.6

# qrcode creator
qrcode==5.3
